FS-1 Changed mixed line endings to CRLF; Moved Player class into Player namespace; Renamed files to match C# references
