+++ Properties/launchSettings.json
@@ -1,0 +1,14 @@
+{
+  "profiles": {
+    "GodotGameWTB": {
+      "commandName": "Project"
+    },
+    "Debug": {
+      "commandName": "Executable",
+      "executablePath": "Godot_v4.1.3-stable_mono_win64.exe",
+      "commandLineArgs": "--path, --verbose",
+      "workingDirectory": ".",
+      "nativeDebugging": true
+    }
+  }
+}
